requirements.txt
pytest
redis
psycopg2-binary